Got device import and fdtd runjob working
- Using Kim et. al Science LatentSpace paper as template

+++ vipdopt/optimization/device.py
@@ -356,7 +356,7 @@ class Device:
 
         # Reinterpolate
         if reinterpolation_factor != 1:
-            new_size = np.ceil(reinterpolation_factor)
+            new_size = np.ceil(reinterpolation_factor).astype(int)
             cur_density_import = repeat(cur_density, (new_size, new_size, new_size))
             # cur_density_import = np.repeat(
             #     np.repeat(
